+++ src/xdoc/mod.rs
@@ -15,8 +15,8 @@ pub use chars::{contains_whitespace, is_whitespace};
 pub use doc::Document;
 pub use doc::{Declaration, Encoding, Version};
 pub use element::Element;
+pub(crate) use name::Name;
 pub use node::{Misc, Node};
-pub use ord_map::OrdMap;
 pub use pi::PI;
 pub use write_ops::{Newline, WriteOpts};
 
@@ -29,8 +29,9 @@ pub mod error;
 mod chars;
 mod doc;
 mod element;
+mod name;
 mod node;
-mod ord_map;
+pub(crate) mod ord_map;
 mod pi;
 mod write_ops;
 
@@ -42,13 +43,7 @@ mod tests {
 
     #[test]
     fn structs_test() {
-        let mut doc = Document::new();
-        doc.set_root(Element {
-            namespace: None,
-            name: "root-element".into(),
-            attributes: Default::default(),
-            nodes: vec![],
-        });
+        let doc = Document::from_root(Element::from_name("root-element"));
         let mut c = Cursor::new(Vec::new());
         let result = doc.write(&mut c);
         assert!(result.is_ok());